requirements.txt
# === Web Backend ===
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
motor==3.7.1
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0

# === Core ===
numpy>=1.24.0
pandas>=2.1.0
torch>=2.1.0
torchvision>=0.16.0
tqdm>=4.66.0
Pillow>=10.0.0

# === Machine Learning & Clustering ===
scikit-learn>=1.3.0
faiss-cpu>=1.7.4

# === Plotting & Visualization ===
matplotlib>=3.8.0

# === Video & Image Processing ===
opencv-python>=4.9.0.80

# === Deep Learning Models ===
open-clip-torch>=2.24.0
transformers>=4.43.0

# === For BLIP-2 (caption generation) ===
accelerate>=0.31.0
safetensors>=0.4.2
sentencepiece>=0.2.0

# === Translate ===
googletrans==4.0.2

# === Misc & Utils ===
requests>=2.31.0

# === Notes ===
# - Compatible with Python >= 3.9
# - Recommended to use a virtual environment (venv or conda)
# - If FAISS GPU errors occur, switch to faiss-cpu
# ====================================

